requirements.txt
ddgs
PyYAML
schedule
pytz
pytest